MessageController: update의 파일 처리에서 버그 있음, 피드백 요청 예정

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -116,7 +116,11 @@ public class BasicMessageService implements MessageService {
                     )
                     .orElse(null);
 
-            message.getAttachments().forEach(binaryContentRepository::deleteById);
+            messageRepository.save(message);
+
+            if(newAttachmentIds != null){
+                message.getAttachments().forEach(binaryContentRepository::deleteById);
+            }
 
             message.update(messageUpdateRequestDto.newContent(), newAttachmentIds);
             return messageResponseMapper.toDto(message);
